% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addPlaylist}
\alias{addPlaylist}
\title{Adds playlist to spotify account}
\usage{
addPlaylist(userid, token, name, uris)
}
\arguments{
\item{userid}{Character string. spotify ID of the user to add the list to}

\item{token}{Character string. access token to the spotify api (has to be on behald of the userid)}

\item{name}{Character string. name of the playlist to be made}

\item{uris}{list of spotify song uris to add to playlist}
}
\description{
Adds playlist to spotify account
}
