% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAccessToken}
\alias{getAccessToken}
\title{Get spotify access token}
\usage{
getAccessToken(code)
}
\arguments{
\item{code}{Character string. spotify authorization code}
}
\description{
Get spotify access token
}
