% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAlbumLyrics}
\alias{getAlbumLyrics}
\title{Return list. contains vector of words in song, data frame of audio features, the title of the album and a
                         list of songs that could not be found}
\usage{
getAlbumLyrics(ID, progressBar = FALSE)
}
\arguments{
\item{ID}{Character. Genius ID of the song}
}
\description{
Return list. contains vector of words in song, data frame of audio features, the title of the album and a
                         list of songs that could not be found
}
