% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getImages}
\alias{getImages}
\title{Return dataframe with urls of artist images}
\usage{
getImages(IDs)
}
\arguments{
\item{IDs}{Character vector. list of spotify artist IDs}
}
\description{
Return dataframe with urls of artist images
}
