% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getNetwork}
\alias{getNetwork}
\title{Return list of nodes and edges based on spotify artist id}
\usage{
getNetwork(ID, progressBar = FALSE, weighted = FALSE)
}
\arguments{
\item{ID}{Character string. artist ID.}

\item{progressBar}{logical that specifies whether a progressbar should be displayed. (only works in shiny)}

\item{weighted}{logical that specifies whether edgelist should be a weighted edgelist}
}
\description{
Return list of nodes and edges based on spotify artist id
}
