% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSongID}
\alias{getSongID}
\title{Return data frame with genius song names and ids}
\usage{
getSongID(query, n = 3)
}
\arguments{
\item{query}{Character string that will be used to search genius}

\item{n}{integer that specifies the number of results to return}
}
\description{
Return data frame with genius song names and ids
}
