% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSongLyrics}
\alias{getSongLyrics}
\title{Return list. contains vector of words in song, data frame of audio features, the title of the song and a
                         name of the artist}
\usage{
getSongLyrics(ID)
}
\arguments{
\item{ID}{Character. Genius ID of the song}
}
\description{
Return list. contains vector of words in song, data frame of audio features, the title of the song and a
                         name of the artist
}
