% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSpotifyID}
\alias{getSpotifyID}
\title{Return data frame with spotify items}
\usage{
getSpotifyID(query, n = 3, type = c("artist", "album", "playlist"))
}
\arguments{
\item{query}{Character string that will be used to search spotify}

\item{n}{integer that specifies the number of results to return}

\item{type}{character string that specifies whether to search for artists, albums or playlists.}
}
\description{
Return data frame with spotify items
}
