% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getUserInfo}
\alias{getUserInfo}
\title{Get current user info}
\usage{
getUserInfo(token)
}
\arguments{
\item{token}{Character string. Spotify access token belonging to current user.}
}
\description{
Get current user info
}
