% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getUserPlaylists}
\alias{getUserPlaylists}
\title{Get list of current user's spotify playlists}
\usage{
getUserPlaylists(userid, token)
}
\arguments{
\item{userid}{Character string. spotify user id of the current user}

\item{token}{Character string. Spotify access token belonging to the current user}
}
\description{
Get list of current user's spotify playlists
}
