% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makePlaylist}
\alias{makePlaylist}
\title{creates recommendation playlist ina tibble that contains spotify tracks.}
\usage{
makePlaylist(playlistID, progressBar = FALSE)
}
\arguments{
\item{progressBar}{Logical indicating whether progressbar should be used (only words in shiny)}

\item{PlaylistID}{Character. spotify ID of the playlist to use as inspiration}
}
\description{
creates recommendation playlist ina tibble that contains spotify tracks.
}
